feat: Profit 생성 메서드 추가

+++ src/main/java/domain/Profit.java
@@ -1,9 +1,9 @@
 package domain;
 
 public class Profit {
-    private final int value;
+    private final double value;
 
-    public Profit(int value) {
+    public Profit(double value) {
         this.value = value;
     }
 }
